+++ src/components/Header.js
@@ -4,7 +4,7 @@ import axios from "axios";
 import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 
-import { selectDocument } from "../feature/editorSlice";
+import { selectDocument, selectDrawingData } from "../feature/editorSlice";
 import {
   changeEditingUser,
   setErrorInfo,
@@ -12,15 +12,19 @@ import {
   selectUserId,
 } from "../feature/userSlice";
 
+import saveCurrentPdf from "../utils/saveDrawing";
+import CONSTANT from "../constants/constant";
+
 export default function Header() {
   const userImage = useSelector(selectUserImgUrl);
   const documentId = useSelector(selectDocument);
   const userId = useSelector(selectUserId);
+  const allDrawingData = useSelector(selectDrawingData);
   const dispatch = useDispatch();
 
   const requestLogout = async () => {
     try {
-      const response = await axios("http://localhost:4000/users/:documents", {
+      const response = await axios(`${CONSTANT.API}/auth/sign-out`, {
         method: "POST",
         responseType: "json",
         withCredentials: true,
@@ -39,34 +43,8 @@ export default function Header() {
     }
   };
 
-  const requestSavePdf = async () => {
-    try {
-      const response = await axios(
-        `http://localhost:4000/users/${userId}/documents/${documentId}`,
-        {
-          method: "PUT",
-          responseType: "blob",
-          withCredentials: true,
-          headers: {
-            Accept: "application/pdf",
-          },
-        },
-      );
-
-      if (response) {
-        const blob = new Blob([response.data], { type: "application/pdf" });
-        const url = URL.createObjectURL(blob);
-
-        const link = document.createElement("a");
-        link.href = url;
-        link.download = "document.pdf";
-        document.body.appendChild(link);
-        link.click();
-        document.body.removeChild(link);
-      }
-    } catch (error) {
-      dispatch(setErrorInfo(error.response.data));
-    }
+  const useSavePdf = () => {
+    saveCurrentPdf(userId, documentId, allDrawingData, CONSTANT);
   };
 
   return (
@@ -75,7 +53,7 @@ export default function Header() {
         <Link to="/">Noted</Link>
       </Logo>
       <NavWrapper>
-        <NavButton onClick={requestSavePdf}>save</NavButton>
+        <NavButton onClick={useSavePdf}>save</NavButton>
         <NavButton>
           <Link to="/">open pdf</Link>
         </NavButton>
