+++ src/constants/constant.js
@@ -1,0 +1,7 @@
+const CONSTANT = {
+  CANVAS_WIDTH: 594.95996,
+  CANVAS_HEIGHT: 841.91998,
+  API: "http://localhost:4000",
+};
+
+export default CONSTANT;
